Consolidated Wiki and Google plugins into Search plugin
