+++ plugins/searchplugin.py
@@ -5,7 +5,7 @@ import urllib.request
 import json
 import re
 
-class WikiPlugin(PlushiePlugin):
+class SearchPlugin(PlushiePlugin):
     name = "Wikipedia Access Plugin"
     description = "Access and summarize English Wikipedia articles"
     authors = ["Garth"]
@@ -87,3 +87,22 @@ class WikiPlugin(PlushiePlugin):
                 bodypart = bodypart.replace("\n", " ")
                 return (True, bodypart, v["title"])
         return (False, "No results found.")
+
+    name = "Google Plugin"
+    description = "Have Plushie search Google for you"
+    authors = ["Kitsune30"]
+    #Shamelessly stolen code from: Garth
+
+    @plushieCmd("google")
+    def googleStuff(self, ctx, msg):
+        args = msg.getArgs()
+
+        if len(args) < 1:
+            ctx.msg("I can't google nothing.", msg.replyTo)
+            return
+
+        if len(args) > 0:
+            url = urllib.parse.quote_plus(msg.noCmdMsg())
+            ctx.msg("Here you go: https://www.google.com/#q={:s}".format(url), msg.replyTo)
+        else:
+            ctx.msg("Something has failed. Please contact Garth about it.", msg.replyTo)
